anv/allocator: Change some parameters and variables from 32bit to 64bits
struct anv_state::offset and struct anv_block_pool::max_size are 64bits so these
parameters should also be 64bit or risk overflow.

Cc: mesa-stable
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37669>
(cherry picked from commit ad62911683793f5b93e1f1f8698e1f38cf4f2a7d)

+++ src/intel/vulkan/anv_private.h
@@ -865,8 +865,7 @@ VkResult anv_block_pool_alloc(struct anv_block_pool *pool,
                               uint32_t block_size,
                               int64_t *offset,
                               uint32_t *padding);
-void* anv_block_pool_map(struct anv_block_pool *pool, int32_t offset, uint32_t
-size);
+void *anv_block_pool_map(struct anv_block_pool *pool, int64_t offset, uint32_t size);
 
 struct anv_state_pool_params {
    const char *name;
